rename: 도메인 객체 패키지 분리
