+++ src/main/java/kitchenpos/domain/menu/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import javax.persistence.Embeddable;
 import java.math.BigDecimal;
