+++ src/main/java/kitchenpos/domain/menu/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import javax.persistence.AttributeOverride;
 import javax.persistence.Column;
