+++ src/main/java/kitchenpos/ui/dto/CreateMenuGroupRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.dto;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menu.MenuGroup;
 
 public class CreateMenuGroupRequest {
