+++ src/main/java/kitchenpos/ui/dto/CreateOrderTableRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.dto;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 
 public class CreateOrderTableRequest {
