+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.repository.MenuGroupRepository;
 import kitchenpos.ui.dto.CreateMenuGroupRequest;
 import org.assertj.core.api.SoftAssertions;
