+++ src/test/java/kitchenpos/domain/MenuProductsTest.java
@@ -1,5 +1,9 @@
 package kitchenpos.domain;
 
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.MenuProducts;
+import kitchenpos.domain.menu.Price;
+import kitchenpos.domain.menu.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
