+++ src/test/java/kitchenpos/domain/OrderTest.java
@@ -1,5 +1,8 @@
 package kitchenpos.domain;
 
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.table.OrderTable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
