+++ src/test/java/kitchenpos/domain/TableGroupTest.java
@@ -1,6 +1,7 @@
 package kitchenpos.domain;
 
-import org.junit.jupiter.api.DisplayName;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
